requirements.txt
numpy==1.23.5
Pillow==10.1.0
PyYAML==6.0.1

